+++ flower.center/flower.center.api/pom.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="UTF-8"?>
 <!--
 
     Copyright © 2019 同程艺龙 (zhihui.li@ly.com)
@@ -20,11 +19,15 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>com.ly.train</groupId>
-        <artifactId>flower</artifactId>
+        <artifactId>flower.center</artifactId>
         <version>0.1.3</version>
-        <relativePath>..</relativePath>
     </parent>
+    <artifactId>flower.center.api</artifactId>
 
-    <artifactId>flower.center</artifactId>
-    <packaging>pom</packaging>
+    <dependencies>
+        <dependency>
+            <groupId>com.ly.train</groupId>
+            <artifactId>flower.common</artifactId>
+        </dependency>
+    </dependencies>
 </project>
