+++ flower.center/flower.center.api/src/main/java/com/ly/train/flower/center/core/store/ServiceConfigStore.java
@@ -1,0 +1,29 @@
+/**
+ * Copyright © 2019 同程艺龙 (zhihui.li@ly.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.ly.train.flower.center.core.store;
+
+import java.util.Set;
+import com.ly.train.flower.common.service.config.ServiceConfig;
+
+/**
+ * @author leeyazhou
+ */
+public interface ServiceConfigStore {
+
+  boolean addServiceConfig(ServiceConfig serviceConfig);
+
+  Set<ServiceConfig> getAllServiceConfig();
+}
