+++ flower.center/flower.center.impl/src/main/assembly/assembly.xml
