+++ flower.center/flower.center.impl/src/main/java/com/ly/train/flower/center/config/FastJsonConfiguration.java
