+++ flower.center/flower.center.impl/src/main/java/com/ly/train/flower/center/core/store/redis/ServiceConfigRedisStore.java
@@ -1,0 +1,57 @@
+/**
+ * Copyright © 2019 同程艺龙 (zhihui.li@ly.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.ly.train.flower.center.core.store.redis;
+
+import java.util.HashSet;
+import java.util.Set;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import com.ly.train.flower.center.core.store.ServiceConfigStore;
+import com.ly.train.flower.common.service.config.ServiceConfig;
+import com.ly.train.flower.common.util.cache.Cache;
+import com.ly.train.flower.common.util.cache.CacheManager;
+
+/**
+ * @author leeyazhou
+ */
+public class ServiceConfigRedisStore implements ServiceConfigStore {
+
+
+  static final Logger logger = LoggerFactory.getLogger(ServiceConfigRedisStore.class);
+
+  private CacheManager cacheManager = CacheManager.get("flower_center_service_config");
+
+
+
+  public boolean addServiceConfig(ServiceConfig serviceConfig) {
+    cacheManager.add(serviceConfig.getFlowName(), serviceConfig, 6000);
+    return true;
+  }
+
+  public Set<ServiceConfig> getAllServiceConfig() {
+    Set<ServiceConfig> ret = new HashSet<ServiceConfig>();
+    Set<String> keys = cacheManager.getAllKey();
+    for (String key : keys) {
+      Cache<Object> cache = cacheManager.getCache(key);
+      if (cache != null && cache.getValue() instanceof ServiceConfig) {
+        ret.add((ServiceConfig) cache.getValue());
+      }
+    }
+    return ret;
+  }
+
+
+}
