--- flower.center/src/main/resources/application.yml
@@ -1,6 +1,0 @@
-spring: 
-   http:
-      encoding:
-         charset: UTF-8
-         enabled: true
-         force: true
