--- flower.center/src/main/resources/flower.yml
@@ -1,4 +1,0 @@
-name: "FlowerCenter"
-basePackage: com.ly.train.flower.center
-host: 127.0.0.1
-port: 8096
