+++ flower.core/pom.xml
@@ -22,9 +22,21 @@
         <groupId>com.ly.train</groupId>
         <artifactId>flower</artifactId>
         <version>0.1.3</version>
-        <relativePath>..</relativePath>
     </parent>
+    <artifactId>flower.core</artifactId>
 
-    <artifactId>flower.center</artifactId>
-    <packaging>pom</packaging>
+    <dependencies>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+        </dependency>
+    </dependencies>
 </project>
