+++ flower.filter/pom.xml
@@ -22,9 +22,25 @@
         <groupId>com.ly.train</groupId>
         <artifactId>flower</artifactId>
         <version>0.1.3</version>
-        <relativePath>..</relativePath>
     </parent>
+    <artifactId>flower.filter</artifactId>
 
-    <artifactId>flower.center</artifactId>
-    <packaging>pom</packaging>
+    <dependencies>
+        <dependency>
+            <groupId>com.ly.train</groupId>
+            <artifactId>flower.common</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.opentracing</groupId>
+            <artifactId>opentracing-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.opentracing.brave</groupId>
+            <artifactId>brave-opentracing</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.zipkin.reporter2</groupId>
+            <artifactId>zipkin-sender-okhttp3</artifactId>
+        </dependency>
+    </dependencies>
 </project>
